refactor(automate-scan): extract config.ts to prepare integration-test fix
