+++ apps/automate-scan/src/index.ts
@@ -1,13 +1,7 @@
 // Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {appRootPath} from '@leav/app-root-path';
-import {loadConfig} from '@leav/config-manager';
-import {IConfig} from 'types';
-import {env} from './env';
 import {startWatch} from './setupWatcher/setupWatcher';
-
-export const getConfig = async (): Promise<IConfig> => loadConfig<IConfig>(appRootPath() + '/config', env);
 
 process.on('uncaughtException', err => {
     console.error('1 - There was an uncaught error', err);
